Python/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
boto3==1.21.13
botocore==1.24.13
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
protobuf==3.19.4
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.2
six==1.16.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
